refactor: make control injection less magic
